refactor(Sections): Section 리스트 일급 컬렉션으로 추출
- 유효성 검사 메서드 추출
