+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -2,23 +2,19 @@ package nextstep.subway.line.domain;
 
 import nextstep.subway.common.BaseEntity;
 import nextstep.subway.section.domain.Section;
+import nextstep.subway.section.domain.Sections;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 public class Line extends BaseEntity {
 
-    protected static final String SECTION_CAN_NOT_BE_NULL = "section null일 수 없습니다.";
-    protected static final String DUPLICATE_SECTION = "이미 등록된 section입니다.";
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -29,8 +25,8 @@ public class Line extends BaseEntity {
 
     private String color;
 
-    @OneToMany(mappedBy = "line", fetch = FetchType.LAZY, cascade = CascadeType.REMOVE, orphanRemoval = true)
-    private final List<Section> sections = new ArrayList<>();
+    @Embedded
+    private final Sections sections = new Sections();
 
     public Line() {
     }
@@ -57,17 +53,11 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public void addSection(Section section) {
-        if (section == null) {
-            throw new IllegalArgumentException(SECTION_CAN_NOT_BE_NULL);
-        }
-        if (sections.contains(section)) {
-            throw new IllegalArgumentException(DUPLICATE_SECTION);
-        }
-        sections.add(section);
+    public Sections getSections() {
+        return sections;
     }
 
-    public List<Section> getSections() {
-        return sections;
+    public List<Section> getSectionValues() {
+        return sections.getValues();
     }
 }
